Allow forcing display controller version
This adds a new `version` option for the `insmod` command when loading the `r61520fb.ko` display controller driver.

Usage: determine the correct version (1-4) by reading `dmesg` output after a successful boot. The line will look something like this:

```
[    1.900000] miyoo panel config as v2
```

Then update `modules.custom.sh` like so:

```
if test -e "./flip"; then
    insmod "./r61520fb.ko" flip=1 version=2
else
    insmod "./r61520fb.ko" version=2
fi
```

If unset, the version will be automatically detected. If set, the set version will be used, but the auto-detection logic will still run and log a warning if it comes up with a different answer.

Setting mine to 1 replicates the behavior I was seeing previously of the display going blank after uboot, but still hearing the "ding".