Set up CI build Makes the config, kernel, devicetree, &  drivers with both the steward and musl toolchains.
I tested and confirmed both work with the mainboot repo.

I also tried the steward/uClibc one with 1.3.3 but it couldn't display anything after the uboot splash. I think it's because the video driver was bundled into the kernel in 1.3.3(?) There's probably a straightforward fix for that, but I'm not sure what it is off the top of my head.

In good news, the file sizes are very close for both builds :)